feat: removed unused pages, moved QrReaderWrapper
