+++ src/renderer/components/common/QrCode/QrReader/style.css
