+++ src/renderer/components/common/QrCode/common/errors.ts
