+++ src/renderer/components/common/QrCode/common/types.ts
