+++ src/renderer/screens/Onboarding/Vault/ManageStepSingle/ManageStepSingle.tsx
@@ -8,7 +8,7 @@ import { Chain } from '@renderer/domain/chain';
 import { ErrorType, SigningType } from '@renderer/domain/shared-kernel';
 import { useChains } from '@renderer/services/network/chainsService';
 import { Button, Input, InputHint, HeaderTitleText, SmallTitleText } from '@renderer/components/ui-redesign';
-import { SeedInfo } from '@renderer/components/common/QrCode/QrReader/common/types';
+import { SeedInfo } from '@renderer/components/common/QrCode/common/types';
 import { useAccount } from '@renderer/services/account/accountService';
 import { createAccount } from '@renderer/domain/account';
 import AccountsList from '@renderer/components/common/AccountsList/AccountsList';
