+++ src/renderer/screens/index.ts
@@ -1,12 +1,9 @@
 import * as Settings from './Settings';
 import * as Staking from './Staking';
 import * as AddressBook from './AddressBook';
-import CameraDev from './CameraDev/CameraDev';
-import ChatDev from './ChatDev/ChatDev';
 import Onboarding from './Onboarding';
 import Operations from './Operations/Operations';
 import Balances from './Balances';
-import Signing from './Signing/SigningDev';
 import Transfer from './Transfer/Transfer';
 import Notifications from './Notifications/Notifications';
 
@@ -17,9 +14,6 @@ export default {
   Operations,
   Onboarding,
   Staking,
-  CameraDev,
-  ChatDev,
-  Signing,
   Transfer,
   Notifications,
 };
